feat: add checklist api - PUT /checklist/{checklistId}, modify checklist api - POST /checklist, DELETE /checklist/{checklistId}

+++ backend/src/main/java/com/lemonmul/gamulgamul/api/ChecklistApi.java
@@ -1,14 +1,15 @@
 package com.lemonmul.gamulgamul.api;
 
-import com.lemonmul.gamulgamul.api.dto.checklist.CategoryDto;
-import com.lemonmul.gamulgamul.api.dto.checklist.ChecklistDto;
-import com.lemonmul.gamulgamul.api.dto.checklist.ChecklistListDto;
-import com.lemonmul.gamulgamul.api.dto.checklist.ChecklistRequestDto;
+import com.lemonmul.gamulgamul.api.dto.checklist.*;
 import com.lemonmul.gamulgamul.entity.checklist.Checklist;
+import com.lemonmul.gamulgamul.entity.checklist.ChecklistBasicItem;
+import com.lemonmul.gamulgamul.entity.checklist.ChecklistCustomItem;
+import com.lemonmul.gamulgamul.entity.product.Product;
 import com.lemonmul.gamulgamul.entity.user.User;
 import com.lemonmul.gamulgamul.security.jwt.JwtTokenProvider;
 import com.lemonmul.gamulgamul.service.CategoryService;
 import com.lemonmul.gamulgamul.service.ChecklistService;
+import com.lemonmul.gamulgamul.service.ProductService;
 import com.lemonmul.gamulgamul.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
@@ -25,6 +26,7 @@ public class ChecklistApi {
     private final ChecklistService checklistService;
     private final UserService userService;
     private final CategoryService categoryService;
+    private final ProductService productService;
 
     /**
      * 체크리스트 리스트 조회
@@ -40,9 +42,9 @@ public class ChecklistApi {
      * 빈 체크리스트 생성
      */
     @PostMapping()
-    public Long createChecklist(@RequestHeader HttpHeaders headers){
+    public ChecklistResponseDto createChecklist(@RequestHeader HttpHeaders headers){
         User user = JwtTokenProvider.getUserFromJwtToken(userService, headers);
-        return checklistService.createChecklist(user);
+        return new ChecklistResponseDto(checklistService.createChecklist(user));
     }
 
     /**
@@ -55,34 +57,85 @@ public class ChecklistApi {
 
     /**
      * 체크리스트 수정
-     * todo 해당 유저의 체크리스트인지 확인 필요?
+     * todo 해당 유저의 체크리스트인지 확인 필요
      */
     @PutMapping("/{checklistId}")
-    public boolean modifyChecklist(@RequestBody ChecklistRequestDto checklistRequestDto){
+    public ChecklistResponseDto modifyChecklist(@RequestHeader HttpHeaders headers, @PathVariable Long checklistId,
+                                                @RequestBody ChecklistRequestDto checklistRequestDto){
+        User user = JwtTokenProvider.getUserFromJwtToken(userService, headers);
+        Checklist checklist = checklistService.checklist(checklistId);
+        checkOwnership(user,checklist);
 
+        fillBasicItem(checklistRequestDto.getChecklistBasicItem(), checklist);
+        fillCustomItem(checklistRequestDto.getChecklistCustomItem(), checklist);
 
-
-        return true;
+        return new ChecklistResponseDto(checklistId);
     }
 
     /**
      * 체크리스트 조회
      * todo 체크리스트 수정 작성 후 동작 확인 필요
-     * todo 해당 유저의 체크리스트인지 확인 필요?
      */
     @GetMapping("/{checklistId}")
-    public ChecklistDto checklist(@PathVariable Long checklistId){
+    public ChecklistDto checklist(@RequestHeader HttpHeaders headers, @PathVariable Long checklistId){
+        User user = JwtTokenProvider.getUserFromJwtToken(userService, headers);
         Checklist checklist = checklistService.checklist(checklistId);
+        checkOwnership(user,checklist);
+
         return new ChecklistDto(checklist);
     }
 
     /**
      * 체크리스트 삭제
-     * todo 해당 유저의 체크리스트인지 확인 필요?
      */
     @DeleteMapping("/{checklistId}")
-    public boolean deleteChecklist(@PathVariable Long checklistId){
+    public ChecklistResponseDto deleteChecklist(@RequestHeader HttpHeaders headers, @PathVariable Long checklistId){
+        User user = JwtTokenProvider.getUserFromJwtToken(userService,headers);
+        Checklist checklist = checklistService.checklist(checklistId);
+        checkOwnership(user, checklist);
+
         checklistService.deleteChecklist(checklistId);
-        return true;
+        return new ChecklistResponseDto(checklistId);
+    }
+
+    /**
+     * 해당 유저의 체크리스트인지 확인
+     */
+    private void checkOwnership(User user, Checklist checklist) {
+        if(!checklist.getUser().equals(user)){
+            //todo 더 적절한 예외 있으면 바꾸기, 아니면 사용자 예외 만들기
+            throw new IllegalArgumentException();
+        }
+    }
+
+    /**
+     * 체크리스트 수정 - 기본 아이템 채우기
+     * todo db에 flush 안되는거 해결하기
+     */
+    private void fillBasicItem(List<ChecklistBasicItemRequestDto> basicItem, Checklist checklist) {
+        List<ChecklistBasicItem> checklistBasicItems = checklist.getChecklistBasicItems();
+        if(!checklistBasicItems.isEmpty()){
+            checklistBasicItems.clear();
+        }
+
+        for (ChecklistBasicItemRequestDto item : basicItem) {
+            Product product = productService.product(item.getProductId());
+            checklistBasicItems.add(ChecklistBasicItem.of(checklist,product));
+        }
+    }
+
+    /**
+     * 체크리스트 수정 - 커스텀 아이템 채우기
+     * todo db에 flush 안되는거 해결하기
+     */
+    private void fillCustomItem(List<ChecklistCustomItemRequestDto> customItem, Checklist checklist) {
+        List<ChecklistCustomItem> checklistCustomItems = checklist.getChecklistCustomItems();
+        if(!checklistCustomItems.isEmpty()){
+            checklistCustomItems.clear();
+        }
+
+        for (ChecklistCustomItemRequestDto item : customItem) {
+            checklistCustomItems.add(ChecklistCustomItem.of(item.getProductName(), checklist));
+        }
     }
 }
